refactor(interface): split api.interface.ts and resolve circular dependencies
Split 1122-line api.interface.ts into three domain-specific files:
- local-client.ts: PlayerAccount, EntitlementsTokenResponse, FriendsResponse,
  PresenceResponse, PresenceJSON
- shared-api.ts: Match types with extracted PreGamePlayer type (duplicated 3x)
- store-api.ts: StorefrontResponse, SkinResponse, WalletResponse, PartyResponse

Move PenaltiesResponse to common.interface.ts to break circular dependency
between common.interface.ts and utils.interface.ts.

Resolves #57
