+++ src/interface/index.ts
@@ -1,3 +1,6 @@
-export * from "./api.interface";
+export * from "./local-client";
+export * from "./shared-api";
+export * from "./store-api";
 export * from "./common.interface";
+export * from "./utils.interface";
 export * from "./valorant-api.interface";
