+++ src/interface/shared-api.ts
@@ -1,109 +1,3 @@
-import { GameState, ProvisioningFlow, QueueId } from "./common.interface";
-
-export interface PlayerAccount {
-	game_name: string;
-	tag_line: string;
-}
-
-export interface EntitlementsTokenResponse {
-	accessToken: string;
-	token: string;
-	subject: string;
-}
-
-export interface HelpResponse {
-	events: { [key: string]: string };
-}
-
-export type FriendsResponse = {
-	friends: {
-		activePlatform: string | null;
-		displayGroup: string;
-		game_name: string;
-		game_tag: string;
-		group: string;
-		/** Milliseconds since epoch */
-		last_online_ts: number | null;
-		name: string;
-		note: string;
-		pid: string;
-		/** Player UUID */
-		puuid: string;
-		region: string;
-	}[];
-};
-
-export type PresenceResponse = {
-	presences: {
-		actor?: unknown | null;
-		basic: string;
-		details?: unknown | null;
-		game_name: string;
-		game_tag: string;
-		location?: unknown | null;
-		msg?: unknown | null;
-		name: string;
-		patchline?: unknown | null;
-		pid: string;
-		platform?: unknown | null;
-		private: string | null;
-		privateJwt?: string | null;
-		presence: PresenceJSON | null;
-		product: "valorant" | "league_of_legends";
-		/** Player UUID */
-		puuid: string;
-		region: string;
-		resource: string;
-		state: "mobile" | "dnd" | "away" | "chat";
-		summary: string;
-		/** Milliseconds since epoch */
-		time: number;
-	}[];
-};
-
-export type PresenceJSON = {
-	isValid: boolean;
-	isIdle: boolean;
-	queueId: QueueId;
-	provisioningFlow: ProvisioningFlow;
-	partyId: string;
-	partySize: number;
-	maxPartySize: number;
-	partyOwnerMatchScoreAllyTeam: number;
-	partyOwnerMatchScoreEnemyTeam: number;
-	matchPresenceData: {
-		sessionLoopState: GameState;
-		matchMap: string | ""; // "/Game/Maps/Juliett/Juliett"
-		provisioningFlow: ProvisioningFlow;
-		queueId: QueueId;
-	};
-	partyPresenceData: {
-		partyId: string;
-		isPartyOwner: boolean;
-		partyState: "DEFAULT" | "MATCHMAKING";
-		partyAccessibility: "CLOSED";
-		partyLFM: boolean;
-		partyClientVersion: string;
-		partyVersion: number;
-		partySize: number;
-		queueEntryTime: string;
-		customGameName: string;
-		customGameTeam: string;
-		maxPartySize: number;
-		partyOwnerMatchMap: string; // "/Game/Maps/Juliett/Juliett"
-		partyOwnerMatchCurrentTeam: string;
-		partyOwnerProvisioningFlow: ProvisioningFlow;
-		partyOwnerSessionLoopState: GameState;
-	};
-	playerPresenceData: {
-		playerCardId: string;
-		playerTitleId: string;
-		preferredLevelBorderId: string;
-		accountLevel: number;
-		competitiveTier: number;
-	};
-};
-
 export type CurrentPreGamePlayerResponse = {
 	Subject: string;
 	MatchID: string;
@@ -112,111 +6,53 @@ export type CurrentPreGamePlayerResponse = {
 
 export type CurrentGamePlayerResponse = CurrentPreGamePlayerResponse;
 
+export type PreGamePlayer = {
+	/** Player UUID */
+	Subject: string;
+	/** Character ID */
+	CharacterID: string;
+	CharacterSelectionState: "" | "selected" | "locked";
+	PregamePlayerState: "joined";
+	CompetitiveTier: number;
+	PlayerIdentity: {
+		/** Player UUID */
+		Subject: string;
+		/** Card ID */
+		PlayerCardID: string;
+		/** Title ID */
+		PlayerTitleID: string;
+		AccountLevel: number;
+		/** Preferred Level Border ID */
+		PreferredLevelBorderID: string | "";
+		Incognito: boolean;
+		HideAccountLevel: boolean;
+	};
+	SeasonalBadgeInfo: {
+		/** Season ID */
+		SeasonID: string | "";
+		NumberOfWins: number;
+		WinsByTier: null;
+		Rank: number;
+		LeaderboardRank: number;
+	};
+	IsCaptain: boolean;
+};
+
 export type CurrentPreGameMatchResponse = {
 	/** Pre-Game Match ID */
 	ID: string;
 	Version: number;
 	Teams: {
 		TeamID: ("Blue" | "Red") | string;
-		Players: {
-			/** Player UUID */
-			Subject: string;
-			/** Character ID */
-			CharacterID: string;
-			CharacterSelectionState: "" | "selected" | "locked";
-			PregamePlayerState: "joined";
-			CompetitiveTier: number;
-			PlayerIdentity: {
-				/** Player UUID */
-				Subject: string;
-				/** Card ID */
-				PlayerCardID: string;
-				/** Title ID */
-				PlayerTitleID: string;
-				AccountLevel: number;
-				/** Preferred Level Border ID */
-				PreferredLevelBorderID: string | "";
-				Incognito: boolean;
-				HideAccountLevel: boolean;
-			};
-			SeasonalBadgeInfo: {
-				/** Season ID */
-				SeasonID: string | "";
-				NumberOfWins: number;
-				WinsByTier: null;
-				Rank: number;
-				LeaderboardRank: number;
-			};
-			IsCaptain: boolean;
-		}[];
+		Players: PreGamePlayer[];
 	}[];
 	AllyTeam: {
 		TeamID: ("Blue" | "Red") | string;
-		Players: {
-			/** Player UUID */
-			Subject: string;
-			/** Character ID */
-			CharacterID: string;
-			CharacterSelectionState: "" | "selected" | "locked";
-			PregamePlayerState: "joined";
-			CompetitiveTier: number;
-			PlayerIdentity: {
-				/** Player UUID */
-				Subject: string;
-				/** Card ID */
-				PlayerCardID: string;
-				/** Title ID */
-				PlayerTitleID: string;
-				AccountLevel: number;
-				/** Preferred Level Border ID */
-				PreferredLevelBorderID: string | "";
-				Incognito: boolean;
-				HideAccountLevel: boolean;
-			};
-			SeasonalBadgeInfo: {
-				/** Season ID */
-				SeasonID: string | "";
-				NumberOfWins: number;
-				WinsByTier: null;
-				Rank: number;
-				LeaderboardRank: number;
-			};
-			IsCaptain: boolean;
-		}[];
+		Players: PreGamePlayer[];
 	} | null;
 	EnemyTeam: {
 		TeamID: ("Blue" | "Red") | string;
-		Players: {
-			/** Player UUID */
-			Subject: string;
-			/** Character ID */
-			CharacterID: string;
-			CharacterSelectionState: "" | "selected" | "locked";
-			PregamePlayerState: "joined";
-			CompetitiveTier: number;
-			PlayerIdentity: {
-				/** Player UUID */
-				Subject: string;
-				/** Card ID */
-				PlayerCardID: string;
-				/** Title ID */
-				PlayerTitleID: string;
-				AccountLevel: number;
-				/** Preferred Level Border ID */
-				PreferredLevelBorderID: string | "";
-				Incognito: boolean;
-				HideAccountLevel: boolean;
-			};
-			SeasonalBadgeInfo: {
-				/** Season ID */
-				SeasonID: string | "";
-				NumberOfWins: number;
-				WinsByTier: null;
-				Rank: number;
-				LeaderboardRank: number;
-			};
-			IsCaptain: boolean;
-		}[];
+		Players: PreGamePlayer[];
 	} | null;
 	ObserverSubjects: unknown[];
 	MatchCoaches: unknown[];
@@ -730,393 +566,4 @@ export type PlayerMMRResponse = {
 	};
 	IsLeaderboardAnonymized: boolean;
 	IsActRankBadgeHidden: boolean;
-};
-
-export type StorefrontResponse = {
-	FeaturedBundle: {
-		Bundle: {
-			/** UUID */
-			ID: string;
-			/** UUID */
-			DataAssetID: string;
-			/** Currency ID */
-			CurrencyID: string;
-			Items: {
-				Item: {
-					/** Item Type ID */
-					ItemTypeID: string;
-					/** Item ID */
-					ItemID: string;
-					Amount: number;
-				};
-				BasePrice: number;
-				/** Currency ID */
-				CurrencyID: string;
-				DiscountPercent: number;
-				DiscountedPrice: number;
-				IsPromoItem: boolean;
-			}[];
-			ItemOffers:
-				| {
-						/** UUID */
-						BundleItemOfferID: string;
-						Offer: {
-							OfferID: string;
-							IsDirectPurchase: boolean;
-							/** Date in ISO 8601 format */
-							StartDate: string;
-							Cost: {
-								[x: string]: number;
-							};
-							Rewards: {
-								/** Item Type ID */
-								ItemTypeID: string;
-								/** Item ID */
-								ItemID: string;
-								Quantity: number;
-							}[];
-						};
-						DiscountPercent: number;
-						DiscountedCost: {
-							[x: string]: number;
-						};
-				  }[]
-				| null;
-			TotalBaseCost: {
-				[x: string]: number;
-			} | null;
-			TotalDiscountedCost: {
-				[x: string]: number;
-			} | null;
-			TotalDiscountPercent: number;
-			DurationRemainingInSeconds: number;
-			WholesaleOnly: boolean;
-		};
-		Bundles: {
-			/** UUID */
-			ID: string;
-			/** UUID */
-			DataAssetID: string;
-			/** Currency ID */
-			CurrencyID: string;
-			Items: {
-				Item: {
-					/** Item Type ID */
-					ItemTypeID: string;
-					/** Item ID */
-					ItemID: string;
-					Amount: number;
-				};
-				BasePrice: number;
-				/** Currency ID */
-				CurrencyID: string;
-				DiscountPercent: number;
-				DiscountedPrice: number;
-				IsPromoItem: boolean;
-			}[];
-			ItemOffers:
-				| {
-						/** UUID */
-						BundleItemOfferID: string;
-						Offer: {
-							OfferID: string;
-							IsDirectPurchase: boolean;
-							/** Date in ISO 8601 format */
-							StartDate: string;
-							Cost: {
-								[x: string]: number;
-							};
-							Rewards: {
-								/** Item Type ID */
-								ItemTypeID: string;
-								/** Item ID */
-								ItemID: string;
-								Quantity: number;
-							}[];
-						};
-						DiscountPercent: number;
-						DiscountedCost: {
-							[x: string]: number;
-						};
-				  }[]
-				| null;
-			TotalBaseCost: {
-				[x: string]: number;
-			} | null;
-			TotalDiscountedCost: {
-				[x: string]: number;
-			} | null;
-			TotalDiscountPercent: number;
-			DurationRemainingInSeconds: number;
-			WholesaleOnly: boolean;
-		}[];
-		BundleRemainingDurationInSeconds: number;
-	};
-	SkinsPanelLayout: {
-		SingleItemOffers: string[];
-		SingleItemStoreOffers: {
-			OfferID: string;
-			IsDirectPurchase: boolean;
-			/** Date in ISO 8601 format */
-			StartDate: string;
-			Cost: {
-				[x: string]: number;
-			};
-			Rewards: {
-				/** Item Type ID */
-				ItemTypeID: string;
-				/** Item ID */
-				ItemID: string;
-				Quantity: number;
-			}[];
-		}[];
-		SingleItemOffersRemainingDurationInSeconds: number;
-	};
-	UpgradeCurrencyStore: {
-		UpgradeCurrencyOffers: {
-			/** UUID */
-			OfferID: string;
-			/** Item ID */
-			StorefrontItemID: string;
-			Offer: {
-				OfferID: string;
-				IsDirectPurchase: boolean;
-				/** Date in ISO 8601 format */
-				StartDate: string;
-				Cost: {
-					[x: string]: number;
-				};
-				Rewards: {
-					/** Item Type ID */
-					ItemTypeID: string;
-					/** Item ID */
-					ItemID: string;
-					Quantity: number;
-				}[];
-			};
-			DiscountedPercent: number;
-		}[];
-	};
-	AccessoryStore: {
-		AccessoryStoreOffers: {
-			Offer: {
-				OfferID: string;
-				IsDirectPurchase: boolean;
-				/** Date in ISO 8601 format */
-				StartDate: string;
-				Cost: {
-					[x: string]: number;
-				};
-				Rewards: {
-					/** Item Type ID */
-					ItemTypeID: string;
-					/** Item ID */
-					ItemID: string;
-					Quantity: number;
-				}[];
-			};
-			/** UUID */
-			ContractID: string;
-		}[];
-		AccessoryStoreRemainingDurationInSeconds: number;
-		/** UUID */
-		StorefrontID: string;
-	};
-	/** Night market */
-	BonusStore: {
-		BonusStoreOffers: {
-			/** UUID */
-			BonusOfferID: string;
-			Offer: {
-				OfferID: string;
-				IsDirectPurchase: boolean;
-				/** Date in ISO 8601 format */
-				StartDate: string;
-				Cost: {
-					[x: string]: number;
-				};
-				Rewards: {
-					/** Item Type ID */
-					ItemTypeID: string;
-					/** Item ID */
-					ItemID: string;
-					Quantity: number;
-				}[];
-			};
-			DiscountPercent: number;
-			DiscountCosts: {
-				[x: string]: number;
-			};
-			IsSeen: boolean;
-		}[];
-		BonusStoreRemainingDurationInSeconds: number;
-	};
-};
-
-export type SkinResponse = {
-	status: number;
-	data: {
-		uuid: string;
-		displayName: string;
-		displayIcon: string;
-		fullTransparentIcon: string;
-		wideArt: string;
-		largeArt: string;
-	};
-};
-
-export type WalletResponse = {
-	Balances: {
-		[x: string]: number;
-	};
-};
-
-export type GameSettingsResponse = {
-	type: "Ares.PlayerSettings";
-	data: string;
-};
-
-export type PenaltiesResponse = {
-	Subject: string;
-	Penalties: {
-		Expiry: string;
-		ID: string;
-		InfractionID: string;
-		IssuingMatchID: string;
-		IsAutomatedDetection: boolean;
-		RiotRestrictionEffect: {
-			RestrictionType: "TEXT_CHAT_MUTED" | "VOICE_CHAT_MUTED" | "PBE_LOGIN_TIME_BAN";
-			RestrictionReason: "INAPPROPRIATE_VOICE" | "INAPPROPRIATE_TEXT";
-		};
-	}[];
-	Version: number;
-	Infractions: {
-		ID: string;
-		Name: string;
-		RatingName: string;
-	}[];
-};
-
-export type PartyResponse = {
-	/** Party ID */
-	ID: string;
-	MUCName: string;
-	VoiceRoomID: string;
-	Version: number;
-	ClientVersion: string;
-	Members: {
-		/** Player UUID */
-		Subject: string;
-		CompetitiveTier: number;
-		PlayerIdentity: {
-			/** Player UUID */
-			Subject: string;
-			/** Card ID */
-			PlayerCardID: string;
-			/** Title ID */
-			PlayerTitleID: string;
-			AccountLevel: number;
-			/** Preferred Level Border ID */
-			PreferredLevelBorderID: string | "";
-			Incognito: boolean;
-			HideAccountLevel: boolean;
-		};
-		SeasonalBadgeInfo: null;
-		IsOwner?: boolean | undefined;
-		QueueEligibleRemainingAccountLevels: number;
-		Pings: {
-			Ping: number;
-			GamePodID: string;
-		}[];
-		IsReady: boolean;
-		IsModerator: boolean;
-		UseBroadcastHUD: boolean;
-		PlatformType: "PC";
-	}[];
-	State: string;
-	PreviousState: string;
-	StateTransitionReason: string;
-	Accessibility: "OPEN" | "CLOSED";
-	CustomGameData: {
-		Settings: {
-			/** Map ID */
-			Map: string;
-			/** Game Mode */
-			Mode: string;
-			UseBots: boolean;
-			GamePod: string;
-			GameRules: {
-				AllowGameModifiers?: string | undefined;
-				IsOvertimeWinByTwo?: string | undefined;
-				PlayOutAllRounds?: string | undefined;
-				SkipMatchHistory?: string | undefined;
-				TournamentMode?: string | undefined;
-			} | null;
-		};
-		Membership: {
-			teamOne:
-				| {
-						/** Player UUID */
-						Subject: string;
-				  }[]
-				| null;
-			teamTwo:
-				| {
-						/** Player UUID */
-						Subject: string;
-				  }[]
-				| null;
-			teamSpectate:
-				| {
-						/** Player UUID */
-						Subject: string;
-				  }[]
-				| null;
-			teamOneCoaches:
-				| {
-						/** Player UUID */
-						Subject: string;
-				  }[]
-				| null;
-			teamTwoCoaches:
-				| {
-						/** Player UUID */
-						Subject: string;
-				  }[]
-				| null;
-		};
-		MaxPartySize: number;
-		AutobalanceEnabled: boolean;
-		AutobalanceMinPlayers: number;
-		HasRecoveryData: boolean;
-	};
-	MatchmakingData: {
-		/** Queue ID */
-		QueueID: string;
-		PreferredGamePods: string[];
-		SkillDisparityRRPenalty: number;
-	};
-	Invites: null;
-	Requests: unknown[];
-	/** Date in ISO 8601 format */
-	QueueEntryTime: string;
-	ErrorNotification: {
-		ErrorType: string;
-		ErroredPlayers:
-			| {
-					/** Player UUID */
-					Subject: string;
-			  }[]
-			| null;
-	};
-	RestrictedSeconds: number;
-	EligibleQueues: string[];
-	QueueIneligibilities: string[];
-	CheatData: {
-		GamePodOverride: string;
-		ForcePostGameProcessing: boolean;
-	};
-	XPBonuses: unknown[];
-	/** Empty string when there is no invite code */
-	InviteCode: string;
 };
